requirements.txt
fastapi
uvicorn
pydantic
python-pptx
requests
